Changing precison with double instead of int

+++ jitp/22.cpp
@@ -11,19 +11,19 @@ int main() {
   for(int i = 0; i < N; i++) scanf("%d", &liczby[i]);
   
   int wynik_elements = ceil(N/2.0);
-  int *wynik = (int*)(calloc(wynik_elements, sizeof(int)));
+  double *wynik = (double*)(calloc(wynik_elements, sizeof(double)));
   
   
   for(int i = 0; i < wynik_elements; i++) {
   
-    if((2*i+1) < N) wynik[i] = (liczby[2*i] + liczby[2*i+1]) / 2;
-    else wynik[i] = liczby[2*i];    
+    if((2*i+1) < N) wynik[i] = ((double)(liczby[2*i]) + liczby[2*i+1]) / 2.0;
+    else wynik[i] = double(liczby[2*i]);    
     
   }
   
   for(int i = 0; i < N; i++) printf("%d ", liczby[i]);
   printf("\n");
-  for(int i = 0; i < wynik_elements; i++) printf("%d ", wynik[i]);
+  for(int i = 0; i < wynik_elements; i++) printf("%.2f ", wynik[i]);
   
   free(liczby);
   free(wynik);
